Published "archive" page, updated the "about us" section in "home"
Removed 1jupp from homepage and put their site to the archive (currently the site is still stored in the "teams" source folder). Their team description will be updated soon.
Updated the text under "about us" on the homepage and added a list of races MATIK's teams have been to. A minor issue with the bullet points being placed on the margins remains.
Added description for team "vurts"